Module Improvements. (#146)
Modern Docking modules work properly now.
