+++ demo-single-app/src/basic/PropertiesDemoPanel.java
@@ -21,10 +21,9 @@ SOFTWARE.
  */
 package basic;
 
-import ModernDocking.AppState;
+import ModernDocking.app.AppState;
 import ModernDocking.DockingProperty;
 import ModernDocking.DockingRegion;
-import ModernDocking.internal.DockableProperties;
 
 import javax.swing.*;
 import javax.swing.text.*;
