+++ demo-single-app/src/packets/PacketListPanel.java
@@ -22,7 +22,7 @@ SOFTWARE.
 package packets;
 
 import ModernDocking.Dockable;
-import ModernDocking.Docking;
+import ModernDocking.app.Docking;
 
 import javax.swing.*;
