+++ docking-api/resources/api_icons/close-16.png
