+++ docking-api/resources/api_icons/close.png
