+++ docking-api/resources/api_icons/settings.png
