+++ docking-api/src/module-info.java
@@ -1,7 +1,7 @@
 /**
  * Module for the Modern Docking framework
  */
-module modern_docking {
+module modern_docking.api {
 	requires java.desktop;
     requires java.logging;
 
@@ -10,6 +10,8 @@ module modern_docking {
 	exports ModernDocking.exception;
 	exports ModernDocking.layouts;
 	exports ModernDocking.persist;
+ exports ModernDocking.settings;
 	exports ModernDocking.ui;
     exports ModernDocking.api;
+	exports ModernDocking.internal;
 }
