+++ docking-multi-app/src/ModernDocking/app/DockableMenuItem.java
@@ -19,8 +19,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-package ModernDocking;
+package ModernDocking.app;
 
+import ModernDocking.Dockable;
 import ModernDocking.api.DockingAPI;
 import ModernDocking.internal.DockingInternal;
