+++ docking-single-app/src/ModernDocking/app/DockingState.java
@@ -19,7 +19,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-package ModernDocking;
+package ModernDocking.app;
 
 import ModernDocking.api.DockingStateAPI;
 import ModernDocking.layouts.ApplicationLayout;
