+++ docking-ui/src/ModernDocking/ext/ui/DockingUI.java
@@ -19,7 +19,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-package docking.ui;
+package ModernDocking.ext.ui;
 
 import ModernDocking.internal.DockedTabbedPanel;
 import ModernDocking.internal.DockingInternal;
@@ -42,7 +42,11 @@ public class DockingUI {
 
         DockingInternal.createHeaderUI = FlatLafHeaderUI::new;
 
-        FlatSVGIcon settingsIcon = new FlatSVGIcon("icons/settings.svg");
+        FlatSVGIcon settingsIcon = new FlatSVGIcon(DockingUI.class.getResource("/ui_ext_icons/settings.svg"));
+
+        if (!settingsIcon.hasFound()) {
+            throw new RuntimeException("settings.svg icon not found");
+        }
 
         DockedTabbedPanel.setSettingsIcon(settingsIcon);
