+++ docking-ui/src/module-info.java
@@ -1,0 +1,10 @@
+/**
+ * Module for the Modern Docking framework
+ */
+module modern_docking.ui_ext {
+	requires modern_docking.api;
+	requires java.desktop;
+	requires com.formdev.flatlaf.extras;
+
+	exports ModernDocking.ext.ui;
+}
